+++ curta/Cargo.toml
@@ -23,7 +23,7 @@ serde = { version = "1.0", features = ["derive"] }
 hex = "0.4.3"
 subtle-encoding = "0.5.1"
 bincode = "1.3.3"
-curve25519-dalek = { git = "https://github.com/succinctlabs/curve25519-dalek.git", branch = "feature/edwards-point-getters" }
+curve25519-dalek = "4.1.1"
 
 [dev-dependencies]
 env_logger = { version = "0.9.0", default-features = false }
