+++ curta/src/chip/ec/edwards/ed25519/mod.rs
@@ -1,4 +1,6 @@
 pub mod decompress;
+pub mod gadget;
 pub mod instruction;
 pub mod params;
+pub mod point;
 pub mod sqrt;
