+++ curta/src/chip/ec/point.rs
@@ -3,22 +3,8 @@ use core::ops::{Add, Neg};
 use num::BigUint;
 use serde::{Deserialize, Serialize};
 
-use super::edwards::ed25519::params::Ed25519BaseField;
 use super::EllipticCurve;
 use crate::chip::field::register::FieldRegister;
-use crate::chip::register::bit::BitRegister;
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
-pub struct CompressedPointRegister {
-    pub sign: BitRegister,
-    pub y: FieldRegister<Ed25519BaseField>,
-}
-
-impl CompressedPointRegister {
-    pub fn new(sign: BitRegister, y: FieldRegister<Ed25519BaseField>) -> Self {
-        Self { sign, y }
-    }
-}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct AffinePoint<E> {
